doc: Add historical documentation
Add the historical documentation that was used to determine
how to actually control the bubble memory module. This was
scattered over the internet, in obscure archives. Without
this documentation, this project would not have been
possible. By gathering it in one place, I hope to help the
next person who decides to play with these devices.

Signed-off-by: evilwombat <evilwombat@peregrine.local>